extract util functions into their own file
